Support importing and returning custom types from modules
Note: If you are reading this because you ran a git blame and this
came up, please know that I had a bad day today and I'm not
thinking clearly.

I relied exclusively on gleam's compiler and unit tests to tell
me I'd done a good job instead of actually reasoning about it.

If you aren't reading this, it means that the compiler is awesome
and the code turned out ok anyway.